requirements.txt
streamlit
pydantic
lance
pandas
daft
pillow
matplotlib
numpy
pyarrow